+++ ak_library_management/__init__.py
@@ -2,3 +2,5 @@
 
 from . import models
 from . import wizards
+from . import report
+from . import controllers